refactor: games.html to use template inheritance and add styles
The games.html template has been refactored to use template inheritance, with the base template being extended via {% extends 'base.html' %}. The <title> tag has been moved to the title block in the base template, and the styles for the page have been added to the styles block. The content of the page remains the same, displaying a list of games in a table format. The scripts block has also been added to the template to include the games.js script. These changes were made to improve the maintainability and organization of the codebase.