refactor: team list HTML to use template inheritance
This commit refactors the HTML code for the team list to use template inheritance. The HTML code has been moved to a child template that extends the base template. The changes include moving the title and the CSS styles to blocks in the child template and adding blocks for content and scripts. This makes the code more maintainable and easier to read.