refactor: team template to extend base template
The team template previously included the HTML head and body elements, as well as a link to the team-specific CSS and JavaScript files. I refactored the template to extend the base HTML template instead, which includes the necessary head and body elements, as well as the base CSS and JavaScript files. This simplifies the code and promotes consistency across the site.