refactor: team creation form to extend base template
This commit refactors the team creation form to extend the base template, which includes the necessary CSS and JavaScript files. The form itself has not changed, but the code has been made more maintainable and easier to read by separating the template into blocks. This change should have no effect on the functionality of the form.