refactor: title page to extend base.html and update form
Changed the HTML code in the "title.html" file to extend the "base.html" file and update the form. The CSS and JS files are still being loaded as before. The form now includes a CSRF token and a class for the "Сохранить" button. The code has been organized into separate blocks for title, styles, content, and scripts.