refactor: teams list page and add CSS and JS files.
This commit refactors the teams list page by moving inline styles and scripts to separate CSS and JS files respectively. The styles have been moved to teams.css and the script has been moved to teams.js. The teams list page has been updated to link to these new files. This makes the code easier to read and maintain, and separates concerns. No changes were made to the functionality of the page.