refactor: delete method to use csrf token and jQuery Ajax
This commit refactors the delete method in the team view to use a csrf token and jQuery Ajax. The csrf token is retrieved from the cookies and included in the request headers before sending the delete request. The jQuery Ajax method is used instead of fetch. If the delete request returns an OK response, the page is reloaded.