refactor: game deletion functionality and add confirmation dialog
This commit updates the game deletion functionality by refactoring the code in 'games.js' to use jQuery and AJAX instead of fetch API. It also adds a confirmation dialog to ask the user if they are sure they want to delete the game before sending the DELETE request. Additionally, it adds the {% csrf_token %} template tag to the delete button in the HTML file to include the CSRF token in the AJAX request headers.