fix: add link to game details page in games list
Added an onclick event to the "Подробнее" button in the games list view. Now, when a user clicks on the button, it redirects them to the details page for that specific game. The link is created using the game's ID and the endpoint /games/game_id/. This change improves user experience by providing a direct link to the details page, instead of forcing the user to search for the game in the list again.