refactor: title.js to use jQuery AJAX instead of fetch
This commit updates the title.js file to use jQuery AJAX instead of fetch to send a PUT request to the server. The code has been updated to use jQuery to get the form and save button, and to use jQuery AJAX to send the request. The CSRF token is now obtained from a hidden input field in the form using jQuery. The code now also includes success and error functions to handle the response from the server.