fix: add delay to menu animation to prevent lagging
The previous version of the code caused the menu to lag when the user quickly moved their cursor over the navigation items. To prevent this, a delay of 100ms was added to the "mouseleave" event to give the menu enough time to close smoothly. The updated code uses a setTimeout() method to delay the execution of the slideUp() function.