refactor: title form and delete button to disable/hide for non-admin users
Changed the title form and delete button to disable or hide for non-admin users. Added a disabled attribute to the title name input field and a hidden attribute to the save button if the user is not staff. Similarly, added a hidden attribute to the delete button if the user is not staff.