refactor: add disabled attribute to input fields and select options if user is not staff.
In the HTML file for displaying detailed information about a game, I added the "disabled" attribute to the input fields for date_of_the_match and main_judge, and to the select options for team1 and team2. This attribute will be added to these fields and options if the user viewing the page is not a staff member. The purpose of this change is to prevent non-staff members from editing this information, as they should not have the ability to make changes to game information.