fix: lagging issue when hovering over navigation menu
Previously, when quickly moving the mouse over the navigation menu, it would cause lagging. This has been fixed by updating the jQuery code to track hover events on the dropdown items of the navbar and show/hide the dropdown menus accordingly. Additionally, we have prevented the menu from closing when clicking on items in the dropdown list to improve user experience.