feat: Add docker-compose-ci.yaml file
This commit adds the docker-compose-ci.yaml file, which contains the configuration for the services required for CI/CD pipeline. The file includes the web service, which runs the Docker image of the web app, the db service, which runs the latest version of the Postgres image and includes the health check, and the redis_db service, which runs the latest Redis image. Additionally, the file includes a common-variables section, which contains the common environment variables used across services.