feat:  Add action.yaml for build and deploy workflow
This commit adds a new file, action.yaml, which contains the configuration for the build and deploy workflow. The workflow is triggered on push events to the master branch and builds and pushes Docker image for the web components of the application. The workflow also deploys the updated application to a remote server using SSH and Docker Compose.