feat: Add nginx service and change static file location
This commit adds a new nginx service to the docker-compose file and changes the location of the static files volume to /usr/src/app/static_files. The nginx service is added to allow serving static files with better performance, and the change in volume location is necessary for nginx to be able to serve the static files. Also, the urls.py file is updated to use the serve() function to serve the static files.