refactor: models and serializers
Changed the models for Title, Team, and Calendar by replacing the foreign key fields "user_id" with "user" for consistency and better readability. In addition, updated the serializers for Title, Team, and Calendar to include a "user" field with a hidden default value of the current user, using the "serializers.HiddenField" and "serializers.CurrentUserDefault" classes. This will allow the backend to automatically set the user field based on the currently authenticated user.