feat: Add redirect to game list after updating a game in game.js and refactor some HTML code in game.html
In game.js, added a redirect to the game list after successfully updating a game. This improves the user experience by bringing the user to the updated game list without requiring an additional action.

In game.html, refactored some HTML code by removing unnecessary default values in form inputs and adding a disabled attribute to the save button for non-staff users. This improves the readability of the code and ensures that non-staff users cannot save changes.