feat: Add rendering of chat list to template
The ChatViewSet class has been updated to include a list method that renders the chat list to the template specified in the template_name attribute. The update method has been removed, and the create method has been updated to use the creator field instead of the user field when creating a new chat object. The functionality of the retrieve method remains the same.